fix: renamed the password files
