+++ src/features/pages/academy-complete/password/password-icon.tsx
