+++ src/features/pages/academy-complete/password/password-input.scss
