+++ src/features/pages/academy-complete/password/password-meter.scss
@@ -43,3 +43,6 @@ $ERROR: #ec3f3f;
         border-radius: 0px 0px 4px 4px;
     }
 }
+.deriv-password__meter {
+    z-index: 2 !important;
+}
