+++ src/features/pages/academy-complete/password/password-utils.ts
@@ -1,4 +1,4 @@
-import { passwordRegex, passwordValues } from './PasswordConstants'
+import { passwordRegex, passwordValues } from './password-constants'
 
 export const validPassword = (value: string) => passwordRegex.isPasswordValid.test(value)
